+++ include/unity/scopes/ScopeMetadata.h
@@ -203,6 +203,16 @@ public:
     */
     int version() const;
 
+    /**
+    \brief Return the list of keywords specified by this scope.
+
+    The list returned by this method is formulated from the value specified under
+    the "Keywords" key in the scope's .ini file.
+
+    \return The list of scope keywords specified by this scope.
+    */
+    std::vector<std::string> keywords() const;
+
 private:
     ScopeMetadata(std::unique_ptr<internal::ScopeMetadataImpl>);           // Instantiable only by ScopeMetadataImpl
     std::unique_ptr<internal::ScopeMetadataImpl> p;