+++ scoperegistry/DirWatcher.cpp
@@ -74,7 +74,7 @@ void DirWatcher::add_watch(std::string const& path)
 
     int wd = inotify_add_watch(fd_, path.c_str(), IN_CREATE | IN_MOVED_TO |
                                                   IN_DELETE | IN_MOVED_FROM |
-                                                  IN_MODIFY | IN_ATTRIB);
+                                                  IN_MODIFY | IN_CLOSE_WRITE);
     if (wd < 0)
     {
         auto msg = "DirWatcher::add_watch(): inotify_add_watch() failed. (fd = " +
@@ -172,9 +172,14 @@ void DirWatcher::watch_thread()
     {
         fd_set fds;
         FD_ZERO(&fds);
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wold-style-cast"
         FD_SET(fd_, &fds);
+#pragma GCC diagnostic pop
 
         int bytes_avail = 0;
+        static_assert(std::alignment_of<char*>::value >= std::alignment_of<struct inotify_event>::value,
+                      "cannot use std::string as buffer for inotify events");
         std::string buffer;
         std::string event_path;
 
@@ -206,12 +211,19 @@ void DirWatcher::watch_thread()
                                        "read() failed on inotify fd (fd = " +
                                        std::to_string(fd_) + ")", errno);
             }
+            if (bytes_read != bytes_avail)
+            {
+                throw ResourceException("DirWatcher::watch_thread(): Thread aborted: "
+                                       "read() returned " + std::to_string(bytes_read) +
+                                       " bytes, expected " + std::to_string(bytes_avail) +
+                                       " bytes (fd = " + std::to_string(fd_) + ")");
+            }
 
             // Process event(s) received
             int i = 0;
             while (i < bytes_read)
             {
-                struct inotify_event* event = (inotify_event*)&buffer[i];
+                auto event = reinterpret_cast<inotify_event const*>(&buffer[i]);
                 {
                     event_path = "";
                     std::lock_guard<std::mutex> lock(mutex_);
@@ -243,16 +255,16 @@ void DirWatcher::watch_thread()
                         watch_event(Removed, File, event_path);
                     }
                 }
-                else if (event->mask & IN_MODIFY || event->mask & IN_ATTRIB)
+                else if (event->mask & IN_MODIFY)
                 {
                     if (event->mask & IN_ISDIR)
                     {
                         watch_event(Modified, Directory, event_path);
                     }
-                    else
-                    {
-                        watch_event(Modified, File, event_path);
-                    }
+                }
+                else if (event->mask & IN_CLOSE_WRITE)
+                {
+                    watch_event(Modified, File, event_path);
                 }
                 i += sizeof(inotify_event) + event->len;
             }