+++ src/scopes/ScopeMetadata.cpp
@@ -157,6 +157,11 @@ int ScopeMetadata::version() const
     return p->version();
 }
 
+std::vector<std::string> ScopeMetadata::keywords() const
+{
+    return p->keywords();
+}
+
 //! @endcond
 
 } // namespace scopes