+++ src/scopes/internal/ScopeConfig.cpp
@@ -57,6 +57,7 @@ namespace
     const string debug_mode_key = "DebugMode";
     const string child_scope_ids_key = "ChildScopes";
     const string version_key = "Version";
+    const string keywords_key = "Keywords";
 
     const string scope_appearance_group = "Appearance";
     const string fg_color_key = "ForegroundColor";
@@ -222,6 +223,22 @@ ScopeConfig::ScopeConfig(string const& configfile) :
 
     try
     {
+        keywords_ = parser()->get_string_array(scope_config_group, keywords_key);
+    }
+    catch (LogicException const&)
+    {
+    }
+
+    for (auto const& keyword: keywords_)
+    {
+        if (keyword.empty())
+        {
+            throw_ex("Invalid empty keyword string found in \"Keywords\" list");
+        }
+    }
+
+    try
+    {
         debug_mode_ = parser()->get_boolean(scope_config_group, debug_mode_key);
     }
     catch (LogicException const&)
@@ -259,7 +276,8 @@ ScopeConfig::ScopeConfig(string const& configfile) :
                results_ttl_key,
                debug_mode_key,
                child_scope_ids_key,
-               version_key
+               version_key,
+               keywords_key
            }
         },
         {  scope_appearance_group,
@@ -441,6 +459,11 @@ std::vector<std::string> ScopeConfig::child_scope_ids() const
 int ScopeConfig::version() const
 {
     return version_;
+}
+
+std::vector<std::string> ScopeConfig::keywords() const
+{
+    return keywords_;
 }
 
 } // namespace internal