+++ src/scopes/internal/ScopeMetadataImpl.cpp
@@ -61,6 +61,7 @@ ScopeMetadataImpl::ScopeMetadataImpl(ScopeMetadataImpl const& other)
     , results_ttl_type_(other.results_ttl_type_)
     , child_scope_ids_(other.child_scope_ids_)
     , version_(other.version_)
+    , keywords_(other.keywords_)
 {
     if (other.art_)
     {
@@ -118,6 +119,7 @@ ScopeMetadataImpl& ScopeMetadataImpl::operator=(ScopeMetadataImpl const& rhs)
         location_data_needed_.reset(rhs.location_data_needed_ ? new bool(*rhs.location_data_needed_) : nullptr);
         child_scope_ids_ = rhs.child_scope_ids_;
         version_ = rhs.version_;
+        keywords_ = rhs.keywords_;
     }
     return *this;
 }
@@ -239,6 +241,11 @@ int ScopeMetadataImpl::version() const
     return version_;
 }
 
+std::vector<std::string> ScopeMetadataImpl::keywords() const
+{
+    return keywords_;
+}
+
 void ScopeMetadataImpl::set_scope_id(std::string const& scope_id)
 {
     scope_id_ = scope_id;
@@ -328,6 +335,11 @@ void ScopeMetadataImpl::set_version(int v)
     version_ = v;
 }
 
+void ScopeMetadataImpl::set_keywords(std::vector<std::string> const& keywords)
+{
+    keywords_ = keywords;
+}
+
 namespace
 {
 
@@ -412,6 +424,15 @@ VariantMap ScopeMetadataImpl::serialize() const
             va.push_back(Variant(sid));
         }
         var["child_scopes"] = Variant(va);
+    }
+    if (keywords_.size())
+    {
+        VariantArray va;
+        for (auto const& keyword: keywords_)
+        {
+            va.push_back(Variant(keyword));
+        }
+        var["keywords"] = Variant(va);
     }
 
     return var;
@@ -555,6 +576,16 @@ void ScopeMetadataImpl::deserialize(VariantMap const& var)
         for (auto const& v: it->second.get_array())
         {
             child_scope_ids_.push_back(v.get_string());
+        }
+    }
+
+    keywords_.clear();
+    it = var.find("keywords");
+    if (it != var.end())
+    {
+        for (auto const& v: it->second.get_array())
+        {
+            keywords_.push_back(v.get_string());
         }
     }
 }