+++ src/scopes/internal/smartscopes/SmartScopesClient.cpp
@@ -290,6 +290,43 @@ bool SmartScopesClient::get_remote_scopes(std::vector<RemoteScope>& remote_scope
             scope.invisible = child_node->has_node("invisible") ? child_node->get_node("invisible")->as_bool() : false;
 
             scope.version = child_node->has_node("version") ? child_node->get_node("version")->as_int() : 0;
+            if (scope.version < 0)
+            {
+                BOOST_LOG_SEV(logger_, Logger::Error)
+                    << "SmartScopesClient.get_remote_scopes(): Scope: \"" << scope.id
+                    << "\" returned a negative \"version\" value";
+                BOOST_LOG_SEV(logger_, Logger::Error)
+                    << "SmartScopesClient.get_remote_scopes(): Skipping scope: \"" << scope.id << "\"";
+                continue;
+            }
+
+            if (child_node->has_node("keywords"))
+            {
+                auto node = child_node->get_node("keywords");
+                if (node->type() == JsonNodeInterface::NodeType::Array)
+                {
+                    auto keywords = node->to_variant().get_array();
+                    for (auto const& keyword : keywords)
+                    {
+                        try
+                        {
+                            scope.keywords.push_back(keyword.get_string());
+                        }
+                        catch (unity::LogicException const& e)
+                        {
+                            BOOST_LOG_SEV(logger_, Logger::Error)
+                                << "SmartScopesClient.get_remote_scopes(): Scope: \"" << scope.id
+                                << "\" returned a non-string keyword";
+                        }
+                    }
+                }
+                else
+                {
+                    BOOST_LOG_SEV(logger_, Logger::Error)
+                        << "SmartScopesClient.get_remote_scopes(): Scope: \"" << scope.id
+                        << "\" returned an invalid value type for \"keywords\"";
+                }
+            }
 
             remote_scopes.push_back(scope);
         }
@@ -698,7 +735,7 @@ std::shared_ptr<DepartmentInfo> SmartScopesClient::parse_departments(JsonNodeInt
         if(subdeps->size() > 0 && dep->subdepartments.size() == 0)
         {
             std::stringstream err;
-            err << "SmartScopesClient::parse_departments(): Failed to parse subdepartments of department '" << dep->label;
+            err << "SmartScopesClient::parse_departments(): Failed to parse subdepartments of department '" << dep->label << "'";
             throw LogicException(err.str());
         }
     }