+++ RELEASE_NOTES.md
@@ -1,10 +1,13 @@
 Release notes
 =============
 
-Changes in version 0.6.8
+Changes in version 0.6.10
+=========================
+  - Renamed "Tags" scope .ini option to "Keywords"
+
+Changes in version 0.6.9
 ========================
   - Added support for ChildScopes scope .ini option to list scopes ids of aggregated scopes.
-  - Replaced dbus-send with "list updated" pub/sub to invalidate smart scopes.
   - Added support for Version attribute in scope.ini file, and added version() accessor
     to ScopeMetadata.
   - Added app_directory() method ScopeBase for scopes that are installed from the same
@@ -12,6 +15,11 @@ Changes in version 0.6.8
     not vice versa).
   - Added missing methods for settings_definitions(), location_data_needed(),
     and child_scope_ids() to testing::ScopeMetadataBuilder.
+  - Added support for Tags scope .ini option.
+
+Changes in version 0.6.8
+========================
+  - Replaced dbus-send with "list updated" pub/sub to invalidate smart scopes.
 
 Changes in version 0.6.7
 ========================